+++ BE/src/main/java/com/codesquad/baseball05/domain/game/dto/ScoreBoardDTO.java
@@ -1,4 +1,4 @@
-package com.codesquad.baseball05.domain.dto;
+package com.codesquad.baseball05.domain.game.dto;
 
 import lombok.Getter;
 import lombok.Setter;
