+++ BE/src/main/java/com/codesquad/baseball05/domain/game/entity/Round.java
@@ -1,4 +1,4 @@
-package com.codesquad.baseball05.domain.entity;
+package com.codesquad.baseball05.domain.game.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
