+++ BE/src/main/java/com/codesquad/baseball05/infra/config/WebConfig.java
@@ -1,12 +1,12 @@
-package com.codesquad.baseball05.config;
+package com.codesquad.baseball05.infra.config;
 
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
+public class WebConfig implements WebMvcConfigurer {
+
     @Override
     public void addCorsMappings(CorsRegistry registry) {
         long MAX_AGE_SECS = 3600;
