+++ BE/src/main/java/com/codesquad/baseball05/infra/config/WebMvcConfig.java
@@ -1,12 +1,11 @@
-package com.codesquad.baseball05.infra;
+package com.codesquad.baseball05.infra.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-public class WebConfig implements WebMvcConfigurer {
-
+public class WebMvcConfig implements WebMvcConfigurer {
     @Override
     public void addCorsMappings(CorsRegistry registry) {
         long MAX_AGE_SECS = 3600;
