+++ BE/src/main/java/com/codesquad/baseball05/infra/dao/MatchDAO.java
@@ -1,6 +1,6 @@
-package com.codesquad.baseball05.infra;
+package com.codesquad.baseball05.infra.dao;
 
-import com.codesquad.baseball05.domain.dto.MatchDTO;
+import com.codesquad.baseball05.domain.matches.dto.MatchDTO;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
