+++ BE/src/main/java/com/codesquad/baseball05/ui/GameController.java
@@ -1,7 +1,7 @@
 package com.codesquad.baseball05.ui;
 
 import com.codesquad.baseball05.application.GameService;
-import com.codesquad.baseball05.domain.dto.ScoreBoardDTO;
+import com.codesquad.baseball05.domain.game.dto.ScoreBoardDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
