+++ BE/src/test/java/com/codesquad/baseball05/application/MatchServiceTest.java
@@ -1,7 +1,7 @@
 package com.codesquad.baseball05.application;
 
-import com.codesquad.baseball05.domain.dto.MatchDTO;
-import com.codesquad.baseball05.infra.MatchDAO;
+import com.codesquad.baseball05.domain.matches.dto.MatchDTO;
+import com.codesquad.baseball05.infra.dao.MatchDAO;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class MatchServiceTest {
